PRE-101 Updates for format changes
Prior to this commit we read the preview reports off disk
This commit changes that behavior to use the new all_nodes key
that was added in PRE-101. The format removed the top_ten key so I now
truncate the all_nodes to manually make the top_ten list as well.

The node breakdown *should* now be sorted by the nodes with the most changes
However I am still researching if this is true. Either way, it now uses the list
of nodes that was provided in the breakdown instead of walking the disk for all
preview reports